+++ src/airflow/model/common/taskinstance.rs
@@ -17,7 +17,7 @@ pub struct TaskInstance {
     pub end_date: Option<OffsetDateTime>,
     pub duration: Option<f64>,
     pub state: Option<String>,
-    pub try_number: i64,
+    pub try_number: u32,
     pub max_tries: i64,
     pub map_index: i64,
     pub hostname: Option<String>,
